Add --name for icingadb in README.md (#64)
* Add --name for icingadb in README.md

The other containers all have meaningful names, describing what they are.
Adding this makes it much easier for people to track the purpose of the container.

* Updated indentation

The for --name argument was indented with spaces, rather than a tab, like the other lines. Correct this.

* Updated indentation

Previous commit missed one --name argument that was indented with spaces, rather than a tab.